requirements.txt
pandas==2.1.1
dash==2.14.2
dash_bootstrap_components==1.5.0
google.cloud
gcsfs
gunicorn